check if hrpsys_tools/hrpsys_ros_bridge is already installed

+++ .travis.sh
@@ -41,15 +41,18 @@ case $TEST_PACKAGE in
         sudo apt-get install -qq -y ros-hydro-$pkg
 
         # this is hotfix
-        sudo wget https://raw.githubusercontent.com/start-jsk/rtmros_common/master/hrpsys_tools/test/test-pa10.test -O /opt/ros/hydro/share/hrpsys_tools/test/test-pa10.test
+        if [ -e /opt/ros/hydro/share/hrpsys_tools ] ; then
+            sudo wget https://raw.githubusercontent.com/start-jsk/rtmros_common/master/hrpsys_tools/test/test-pa10.test -O /opt/ros/hydro/share/hrpsys_tools/test/test-pa10.test
+        fi
 
-        sudo touch /opt/ros/hydro/lib/python2.7/dist-packages/hrpsys_ros_bridge/__init__.py;
+        if [ -e /opt/ros/hydro/lib/python2.7/dist-packages/hrpsys_ros_bridge/ ] ; then
+            sudo touch /opt/ros/hydro/lib/python2.7/dist-packages/hrpsys_ros_bridge/__init__.py;
 
-        #
-        sed -i s@imu_floor@odom@g /opt/ros/hydro/share/hrpsys_ros_bridge/test/test-samplerobot.py
+            #
+            sed -i s@imu_floor@odom@g /opt/ros/hydro/share/hrpsys_ros_bridge/test/test-samplerobot.py
 
-        #
-        patch -p0 /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py <<EOF
+            #
+            patch -p0 /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py <<EOF
 --- /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py 2014-04-17 17:28:42.000000000 +0900
 +++ /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py 2014-04-28 00:30:27.250839313 +0900
 @@ -21,7 +21,7 @@
@@ -64,7 +67,7 @@ case $TEST_PACKAGE in
 
 EOF
 
-        patch -p0 /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch <<EOF
+            patch -p0 /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch <<EOF
 --- /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch 2014-04-17 17:28:42.000000000 +0900
 +++ /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch 2014-04-28 00:30:27.250839313 +0900
 @@ -10,9 +10,9 @@
@@ -80,6 +83,7 @@ EOF
    <arg name="USE_TORQUECONTROLLER" default="false" />
    <arg name="USE_SOFTERRORLIMIT" default="true" />
 EOF
+        fi
         #
         source /opt/ros/hydro/setup.bash
